Implement key algorithm detection in danebot, to replace setting it for all keys. (#2)
_This PR presupposes the logging PR to be merged, as it logs warnings about fallbacks._

#==================

### Added key algorithm detection for renewal configuration.

This change removes the original setting of a default key type and key size, which forced the user to set environment variables before some invocations, in case the key types and/or sizes/curves are different from that default setting.

The new function retrieves the key type and - if available - key size / curve from certbot's renewal configuration file.

In case the size or curve are missing, a warning is logged, and certbot's default settings (for RSA: 2048, for ECDSA: 256) are used as a fallback.

In case you are using a really old version of certbot (< 1.xx) on an old OS (e.g. CentOS 7, Debian 10, Ubuntu 18.04), where only RSA was available, and therefore no key type was specified, the fallback is RSA 2048 (certbot's defaults at that time).